smurf: fts2gaia: Display in-band FTS-2 spectrum in gaia
Display those frames of an FTS-2 spectrum which fall within the band-pass.
For 850 um, this band is between 11.2 and 12.2 cm^-1 wave numbers, and
for 450 um, it is between 22.1 and 23.3 cm^-1 wave numbers.
This translates to the peak of the spectrum of the target,
found near the middle of the sensor array plus and minus some number of
frames corresponding to the desired portion of the band pass.