smurf: Allow user to say what to do next if makemap is interupted.
After a control-C, the current iteration will complete and the user will
then be prompted (using parameter INOPTION) for what to do next. Options
include 1) abort immediately with an error, 2) close immediately retaining
the current potentially unfinalised output map, 3) perform one more
iteration to finalise the current map and then close.

An extra iteration is needed if any AST masking is used since a fnal
iteration must be done with no masking to get the correct flux for the
areas outside the mask.