smurf: Allow itermaps to be written to a different output file.
Specifying a value for ADAM parameter "ITERMAPS" will cause itermaps to
be placed in the specified file rather than in the main output NDF. This
is good because it means you can look at earlier itgermaps whilst makemap
is still running. If makemap aborts for any reason (computer crashes or
whatever), the itermaps will still be usable (unless you are very unlucky
and the crash happens whilst an itermap is being written), allowing you
to use than as the initial sky on a subsequent run of makemap.