smurf: Allow the AST model to be skipped on initial iterations
The new config parameter "ast.skip" specified the number of iterations
(default is zero) for which no AST signal will be subtracted from the
residuals. A map is still created at the end of each iteration, but no
AST signal is derived from it and the residuals are left unchanged.

This means that each of these iteration will start from essentially the
same residuals (i.e. the initial cleaned raw data). However, the creation
of the map allows an FLT mask based on SNR values (flt.zero_snr etc) to be
created.