ast: New sorting options for KeyMap
The existing AgeUp and AgeDown SortBy options use the age of the value as
the sort key. This means that changing the value associated with an
existing entry will change the position of the entry in the sort order.

The new KeyAgeUp and KeyAgeDown SortBy options use the age of the key
as the sort key. This means that changing the value associated with an
existing entry will NOT change the position of the entry in the sort order.