smurf: Return even if cache was not freed
Fixes #3

sc2filtermap does not work on time-series data so the WVM cache was
never created. The previous code was:

    if (freeing_cache) {
        if (CACHE) {
          CACHE = astAnnul( CACHE );
          return VAL__BADD;
        }
    }

so the code freeing the cache called at the end of sc2filtermap in
smurf_mon.c did not actually return and immediately went into the
sanity check logic to check whether hdr was defined. Obviously it
was not.

Now always return after checking whether the cache needs to be
freed.