docs: Move SUN/93 to obsolete.
Being a document about TeX, Jo wrote it in TeX not LaTeX.  The regular
build assumed that it was LaTeX and hence the nightly Jenkins build
failed.  While we could change the build to run tex rather than latex,
there doesn't seem much point.

The Starlink standard has been LaTeX using template prologues for 25
years, to give a uniformity of style (for example see SGP/28).  Also
the standard typesetting system in astronomy has been and continues
to be LaTeX.  So there's little point in having an introduction to
TeX.