smurf: Ignore errors if the cache free fails
There are reports of CALCQU triggering errors during the cleanup
of the cache. I am guessing that CALCQU does not attempt to fix
up header metadata so does not initialise an early AST object in
the main thread. Later on the WVM data are calculated in threads
such that when the object is freed in the main thread at task
exit the object is invalid even though the pointer is defined.

The error message is:

!! AST: Error at line 160 in file smf_calc_wvm.c.
!  Invalid Object pointer given (value is 390951).
!  This pointer has been annulled, or the associated Object deleted.
!  Application exit status AST__OBJIN, Object invalid

which hints that my interpretation may not be the whole story.