smurf: Translate GSD frontend string to the real instrument name
It turns out that GSD frontend strings are not very consistent
with the worst offender being the switch of RXB3CU to RXB instead
of just using RXB3.

This mainly affects the B-band instruments and RxW now gets the
band appended to the instrument name to be consistent with ACSIS
usage.

A full table of instrument history is included in the routine
gsdac_getRealInstrumentName.c