cfitsio: Install versioned shared libraries
CFITSIO now has version libraries. This is very good but we have
to install them and it does not happen automatically. For now
the libraries are explicitly listed in the Makefile.am but
the installer does not know they are soft links so we get three
copies of the libraries rather than the original softlinks pointing
to every more specific versions.