smurf: relax flatfield test for incorrect gradient of poly fit
Previously, a bolometer was flagged as bad if the fast-flat observed
current versus heater power had a temporary downwards wiggle at the
central heater power value. This is because the central value was used a
a reference value. This can happen a lot with noisey fast-flat data such
as is generated by POL-2 when the calibrator is in the beam, resulting in
all (or most) bolometers being rejected.

Now, if the central current value is unusable due to a wiggle, a nearby
good value is used as the reference value instead. This allows more
bolometers to get through to the SNR test, so setting a low value
for parameter FLATSNR  will then allow pol-2 data to be flat-fielded.