kappa: Switch to pdflatex
Remove sun95_*.gif files as they were poor copies of the screen grabs
and the embedded PDFs are better. In theory we could simply convert
the screen grab PDFs to PNG but I will save that for later. In future
we should just screen grab in PNG and use that natively for the PDF.

Converted example GIF files to PNG.

Latex file no longer has special section for GIF.