kaplibs: Fix bug updating input argument.
This didn't matter until Commmit de25d70cb6b60, where NGOOD could be
changed by excluding bad pixels, and is a consequence of extracting
from stats.f into a subroutine.  A further ramification is it is safer
to determine the indices of the percentiles _after_ shifting the index
array for all the bad values in KPG1_MVBDx.