kappa: Allow normalize to loop if IN1 is 1D and IN2 is 2D
If IN1 spans a single row or column, and IN2 spans many rows or columns,
and LOOP=YES, then IN1 will be normalised multiple times, using each row
or column of IN2 in turn to define the normalisation. In this case, the
output NDF will have the shape and size of IN2, rather than IN1 as
usually happens.

Prior to this change, looping could only be used if IN1 was 2D and IN2
was 1D