star2html: Convert PDF images to PS and copy PNG files
latex2html can not handle PDF embedded images and it will be quite
a hack to fix this. Currently latex2html runs latex so that all
the includegraphics options are processed properly. It then converts
the DVI file to postscript files (one file per "page") and uses
pstoimg (ghostscript and netpbm) to convert the output PS files
to PNG images. Converting this to pdflatex and adjusting pstoimg
to convert the PDF pages to images is not a completely crazy job
but it seems a big enough change that it would make the Starlink
copy of latex2html completely different and there is no obvious
place to send the patches upstream.

For now star2html will scan the tex file for includegraphics lines
and determine if any PDF files are being embedded. If they are it
runs pdf2ps and converts them to postscript so that the standard
mechanism will work. Star2html will complain if the
includegraphics directive includes an explicit pdf suffix as that
will break the conversion within latex2html.

latex2html can locate PNG files in includegraphics lines so long
as the files are in the current directory. Star2html does the conversion
in a temporary directory and assumes that images will be noted
explicitly with \htmladdimg and mentions in .htx_tar.extras files. This
means that includegraphics will not work for PNG. To overcome this
star2html now scans the tex file and copies relevant PNG files into
the working directory.