figaro: Reinstate the GIF files for the HTML version
When we integrated Figaro into CVS in 2004 it seems we missed the
GIF images of the screen grabs in the addon/ directory. Malcolm
Currie located the addon directory in a 2003 Starlink CD and they
are now included.

Use the .extras scheme to copy the files into the htx_tar directory.
This required that we removed the addon/ part of the path in the
tex file that references the images with htmladdimg.