KAPPA: Remove loitering NDF_UNMAP.
The unmapping for making the residuals map (obviously not commonly
used) was moved earlier to free resources as soon as they could be
(178b158f65be), making a later unmap incorrect.  This change removes
that map.  In the meantime I had added blocking (7b9ec8b0f960b), so
the INNDF identifier's data that MFITTREND was attempting to unmap is
now never mapped; instead it's the input block identifier IBL that's
free'd earlier.