kappa: Allow mfittrend to preserve the fit where the input is bad
If SUBTRACT is FALSE (i.e. the output NDF is the fit itself rather than
the background-subtracted data), then the new parameter PROPBAD can be
used to determine if the fit should be set bad if the corresponding input
pixels are bad. The default value is TRUE, which retains previous
behaviour.