smurf: Updated list of nights WVM was unusable for smurf.
After review, the night 20130422 was found not to require the use of CSO
fits. It showed good (indeed, excellent) WVM response for nearly the entire
night, and the only reason I included it on the list was because of a period
about and hour long where it seemed to be behaving erratically. This period
was separate from periods of good performance both before and after it, and
was during a time when the CSO WVM reported a high amount of water vapor in
the atmopsphere, and the WVM is known to occasionally behave sporadically
in inclement weather.

Moreve, and more importantly, no data was taken during that period, and since
the rest of the night the WVM was behaving itself fine, I felt it should
probably be removed from the list of nights where it was acting up.