smurf: Ensure correct wcsalign params are used in pol2cat
The default value for REBIN is sticky and so we need to ensure the
correct value is used each time wcsalign is run. Also, the default value
for CONSERVE depends on the value of REBIN. When aligning Q and U values
with te reference image, REBIN=yes was used, which means CONSERVE=yes was
being used by default. This was the wrong thing since Q adn U values
represent *average* values, not sums. So set CONSERVE=NO in these cases.