smurf: Allow skyloop to use noi.box_size
This involved changing from using noi.calcfirst=1 to noi.calcfirst=0. The
NOI model is now calculated after the first iteration is completed, and
is exported for use on subsequent iterations. Previously, it was
calculated afresh on every iteration, before the iteration commenced.
This means that it is now possible for skyloop to recognise and use the
noi.box_size parameter. But first results suggest that the noise values
are less stable without calcfirst set, causing some bolometers to be
given inappropriately small variances, and thus be over-emphasised in the
final map, resulting in visible bolometer tracks. If this is a problem,
add "noi.calcfirst=1" to your config., and remove "noi.box_size".