ast: Improve handling of headers with SIP distortion
- Since the inverse polynomials supplied in SIP headers often seem to be
very inaccurate, always ignore them and re-generate inverse polynomials
from the supplied forward polynomial.

- Only use the "fit a polymap to the inverse SIP transformation" option
if the fit is accurate to 0.01 pixel over an area 3 times the dimensions
of the FITS file. This is a stricter requirement than previously since
people want to be able to do accurate transforms outside the image area.
If the fit cannot be done over such a large area, then tell the PolyMap to
use a Newton-Raphson method to invert each point indivudally (this is slower
but more accurate than using a single fit for all points).