smurf: Provide filter to flag ringing after FLT
Setting FLT.RING_BOX1 to a value such as 0.5 causes a filter to be run on
the residuals after subtraction of FLT. This filter looks for and flags
samples that appear to suffer from ringing (i.e. have a large amplitude
oscillation near a frequency close to the FLT edge frequency). These
samples are excluded from the map.

To avoid problems with convergence, the flags are cleared before starting
the next iteration (i.e. when the FLT is added back onto the residuals
prior to COM). This means the flagging process does not influence the COM
model.