smurf: Tidy FTS-2 related code in sc2ast
Commit 19cdf0f85cb4eac3ff2b1e31adc779b22998b7c3 corrected the naming
of the ports so that they were referred to as either TRACKING or
IMAGE rather than by the physical port.  That change was lost in
2c63e3c51b919dd254a948c4566220904ce9c99f and so this commit aims
to restore the TRACKING and IMAGE names for the variants of the
WCS information.

The code has also been simplified as there are only the two options
(TRACKING and IMAGE) and the order in which they are added probably
does not matter. The TRACKING variant is now always added first,
then the IMAGE variant.  The TRACKING variant is then re-selected
if the user has requested that their map be made using the TRACKING
coordinates.