smurf: Index JSA tile facets as defined by HEALPix
The HEALPix paper (Gorsky et. al. 2005 ApJ 622, 759) defines the
numbering scheme to be used for the base resolution pixels (facets).
This commit adjusts the numbering of the facets to follow this
scheme.  This is the first step towards numbering the JSA
tiles using the HEALPix nested numbering scheme, which was
an action item from the JSA meeting of 2013-10-28.