smurf: Nested indexing of JSA tiles within facets
The tiles within each HEALPix base resolution pixel (facet)
are now numbered using the nested scheme described in the
HEALPix paper.  The offsets north-east and north-west from
the southern corner of the facet are represented by the
even (including 0th) and odd bits respectively.  This,
together with the previous commit, completes the action
item from the JSA meeting on 2013-10-28 of applying the
standard HEALPix nested numbering scheme to the JSA tiles.