smurf: Check for -ve tile coords before division
In smf_jsatilexy2i the x and y indices of the tile in the
all-sky map must be checked before dividing by the number
of pixels per side of a facet.  Otherwise small negative
numbers become zero upon integer division, and coordinates
within one facet below or to the left of the valid area
are not rejected, leading to the generation of spurious
tile indices.