smurf: makemap can now split the output map into JSA tiles
Set the new parameter JSATILES to TRUE to cause the output map to be
split into JSA tiles. Each tile has the name given by OUT but with the
tile index appended. The NTILE output parameter is set to indicate the
number of tiles created, and the tile NDFs are written to the file
specified by OUTFILES. Each output NDF has a FITS header "JSATILE"
holding the index of the tile. The "JSATILE" parameter over-rides the
"REF" parameter.