CUPID: Improve the UI of OUTLINECLUMP.
Jan had reported difficulties using OUTLINECLUMP.  One was because his
NDF was called y.  Adding the @ prefix did not work because the script
treated the @ literally.  While looking at this other deficiencies soon
became apparent.  The prompt for INDEX did not recognise a ? response.
There was no validation of the input parameters if they were given on
the command line.  On exit the GLOBAL.DATA_ARRAY association was
changing to the model, rather than the NDF supplied as the user might
expect.  There was no indication of what was the maximum index, which
could lead to a confusing second prompt for NDF when it was the INDEX
that was wrong.

So I've restructured the code and tackled the known problems.

There is a clunky use of the output from HDSTOOLS::HDIR to obtain the
number of indices.  HDIR ought to write out the key values to
parameters to assist scripting.  Another issue is that SUN/255 has not
release notes and It's ostensibly at the 0.0 version instead of 1.6.