CONVERT: Add a parameter to permit export of tabular WCS via the TAB algorithm.
We apologise for the late running of this commit.  This change arose
out of Tim's desire to demonstrate our sparse data to Doug Tody under
the thread "Sparse cubes and FITS" back in March.  At that time I
wasn't sure that a decision for the way ahead had been made.

The USEAXIS parameter was being overworked.  This introduces a
separate parameter called ALLOWTAB (defaulting to TRUE) to control
whether or not tabular co-ordinates can be written using the TAB
algorithm.  Remove a redundant IF block, instead just not set the AST
TabOK attribute when ALLOWTAB is set FALSE.