convert: Inherit WCS from FITS extension if none present in NDF
An unfortuante side-effect of the previous commit (eb8f579a48f) is that
when exporting an NDFthat 1) has valid WCS keywords in its FITS extension
but 2) has no WCS coomponent, the WCS in the FITS extension is no longer
used to create WCS keywords in the exported file.

This commit fixes that issue.