CUPID: Permit more than one clump to be plotted.
This was a request from Jan who found it tedious to plot each outline
individually.

It could be more clever with max and min "-x" to go from 1 to x
and x-" to plot index x until the last index.  At present both will
display indices 1 to x, as a index list like "5-3" is equivalent to
"3-5".