kappa: Normalise the formatted axis values written out by centroid
Since the CENTRE, XCEN and YCEN parameters hold *formatted* axis values,
they should be normalised. The rule is effectively, always normalise
before displaying or passing on formatted values, but no need to normalise
if displaying or passing on unformatted values.

This is because the purpose of formatting is to produce human-readable
values.