smurf: Use slow mode for finding map bounds when using JSA pixel grid
The fast mode in smf_mapbounds assumes that pixels are square, which is
not the case for HPX projections. Consequently, the wrong extreme time
slices were being found and used, resulting the final map being a bit too
small. Now, slow mode is used when creating a map on the JSA pixel grid.