smurf: Use FITS-WCS instead of FITS-AIPS to specify map rotation.
Instead of using the specified rotation angle (parameter CROTA) to set
the FITS-AIPS "CROTA2" keyword for a TAN projection, use it to set the
FITS-WCS "PCi_j" keywords. In the FITS-AIPS scheme the pixel dimensions
are meaured along the X and Y pixel axes. In the FITS-WCS scheme they
are measured along the celestial north and east axes. Put another way,
non-square pixels are always elongated along a pixel axis in the old
scheme, and are always elongated along a celestial axis in the new
scheme. The two schemes are equivalent in the normal (universal?) case
of square pixels or if CROTA is zero. If the pixels are not square and
CROTA is not zero, they produce different results.