figaro: Use PSX_GETCWD and remove GEN_GETCWD
GEN_GETCWD indicated that it was hoping that PSX_GETCWD would
arrive some time in 1993. Interestingly the change log
for one of the DSA routines indicates that it did exist around
that time and was reverted. This is odd because the PSX in the
repository did not get PSX_GETCWD until commit
415ea5890df8dff831611f36836597580d01fd8e

GEN_GETCWD had compiler errors because it was using a deprecated
BSD API with getwd(). Seemed simpler to replace it with PSX.