smurf: Fix bug that cuased first contchunk to get double weight
Before e93e76879, smf_iteratemap used the returned map array to hold the
first chunk, and allocated memory for any remaining chunks. smf_addmap1
used this fact to identify when the first chunk was being processed [OK -
I know - a very bad idea]. Since e93e76879, the first chunk is now stored
in allocated memory just like all other chunks. This caused smf_addmap1
to treat the first chunk like all the other chunks, ignoring the fact
that the first chunk is already present in the supplied arrays. The
upshot of this was that the first chunk got included twice in the
returned map, effectively giving it double weight.

smf_addmap1 now has an extra parameter to indicate if the first conchunk
is being processed.