cupid: Prevent logfile columns merging together
Now explicity include a space between the columns of the CUPID
log file.  Previously certain numbers such as -0.00049505531414
could overflow into the adjacent column which would then continue
with no separation.  The problem with the formatting expression
"%-*.*g", LOGTAB, LOGTAB-5 is that the second number is the
number of significant digits, but the example number above
has 6 characters before the first significant digit, so the
restriction doesn't prevent the column overflowing.  Since the
first formatting parameter specifies the minimum width,
the column can still be too wide, causing misalignment
of the table.  Lines in the table are also one character
shorter, but another space could be added to the end of
every line if that turns out to be a problem.