perlmods: Add UKIRT::Archive module
This module contains the "ukirtwrapdr" script and associated
library file.  This is intended to be an equivalent of
"jsawrapdr" for the UKIRT archive at CADC.  This commit
also updates the JSA module to the (refactored) version
required by UKIRT::Archive.

Note that this version only includes dummy expressions
for UKIRT data product filenames.