atools: Allow astoutline to create poly in current NDF frame
New parameter CURRENT can be set TRUE to cause the polygon returned
by astOutline to be mapped into the current Frame of the NDF (using
astMapRegion). Defaults to FALSE in order to retain a one-to-one
correspondance between the default behaviour of the atools "astoutline"
command and the AST "astOutline" function.