smurf:: Enable skyloop to do map-based despiking
SKYLOOP was not doing any map-based despiking because the despiking was
not being done on the first iteration of makemap due to the NOI model not
being available. But in SKYLOOP every iteration is the first iteration as
far as makemap is converned, and so no despiking was being done. In fact
when running smakemap from within SKYLOOP, the NOI model *is* available
on the first iteration because it will have been imported from a disk
file created by an earlier invocation of makemap. The problem is just in
the way that smf_calcmodel_ast determines whether NOI values are
available or not. So I've changed this to use the first value in the NOI
model as the flag - if this value is 1.0 (the initial value before any
NOI values have been calculated), then no despiking is done. If the first
NOI value is not 1.0, then this is used as a flag that NOI values are
available and despiking is therefore done.