SMURF: Revert d373bf96cbb6467a767ad.
I thought this was ready but I had forgotten after returning from
Victoria that it hadn't compiled, and I was trying to get the
modification out of the way so that I could do a git pull while git
stash had a problem.  The case changing still needs to be done, but
implemented so the code compiles.