smurf: Use variances from penultimate iteration if flt.filt_edge_largescale_last set
Using a different filter on the last iteration changes the map variances
in the source region but does not change the map data values (much). To
avoid this change, the map variances calculated on the penultimate
iteration are now stored in the final map if flt.filt_edge_largescale is
set.