kaplibs: Allow KAPLIBS to read regions from the outline extension of an NDF
KPG1_GTOBJ (used for instance by kappa:ardplot and all the atools
commnds) will now read and return a polygon from the OUTLINE extension of
an NDF, if present. It uses the new kpgGetOutline function to do this. So
for instance

% kappa
% ardplot region=myndf

will plot the STC-S polygon in myndf.sdf (if one exists).