ast: Provide function to generate convex hull polygon
A convex hull is the shortest polygon that encloses all selected
positions, so if the good pixels in a map are in more than one contiguous
lump (as can be the case with ACSIS data), a convex hull is probably
better as a description that a single polygon giving the outline of a
randomly chosen lump.

See function astConvex in polygon.c.