kaplibs: Do not specifiy uncertainty in convex hull
The recent changes to AST regarding simplification of Polygons means that
it is no longer necessary to specify an uncertainty in the outline
polygon. Previously, bending of the polygon edges caused by the
non-linear mapping from pixel coords to sky coords caused the
simplification of the Polygon to fail, and so prevent it being converted
to STC-S format.