smurf: Allow sc2concat output NDFs to be specified by base name
The new parameter OUTBASE allows the names of the output NDFs to be
generated automatically as "<outbase>_<n>.sdf", for <n> = 1, 2, 3, etc.
This means you do not need to know how many output NDFs will be created,
before running sc2concat.

The default is still to use the old OUT parameter, which requires that
you know how many output NDFs are going to be created.