smurf: Fix problem with COM/FLT masking in skyloop
COM and FLT masks cannot be created until a map is available. Previously,
this test was based on the iteration number, on the assumption that no
map would be available until the end of the first iteration. However,
this is not a good check since a map may have been provided via parameter
REF (as in skyloop) and so be avilable before the end of the first
iteration. This is important since makemap only ever does one iteration
when called from within skyloop.

So change makemap to use an explicit flag that indicates whether a map is
available or not.