kappa: Allow ardplot to plot over an empty display
Previously, ardplot required that another picture be displayed first,
with which the Region to be plotted could be aligned. Now, it can plot
the Region on an empty display. To do this it creates a basic TAN
projection (for sky regions), mapped onto the display surface so that the
Region is centred on the display. The size of the plot can be controlled
by the new SIZE parameter.