smurf: All source pixels to be accumulated within an SNR mask
Using an SNR mask can introduce instabilities in which pixels close to
the SNR limit swap repeatedly from inside to outside the mask, thus
preventing convergence. Setting the new (AST|COM|FLT).ZERO_ACCUM
parameter to a non-zero value prevents this by accumulating the mask from
one iteration to the next. In other words, if a pixel is flagged as
source, it remains flagged as source for all further iterations -
"Once a source pixel, always a source pixel".