smurf: Update quality in initial sky map
When makemap is used on multiple chunks of data, the quality array in the
output map is a logical OR of the maps for the individual chunks. This is
not very useful for skyloop, so now the SNR-based masked actually used is
written back out to the initial sky map when the initial sky is first accessed.