smurf: Allow map to be lagged between iterations
On each iteration, the used map is a weighted mean of the new map and the
map from the previous iteration. The "maplag" parameter (default is zero)
gives the weight for the previous map. The weight for the new map is
(1-maplag). A non-zero value for maplag may help to prevent oscillations
in the SNR mask that can sometimes prevent convergence.