smurf: Prevent mapchange being dominated by bad edge pixels
Pixels round the edge of the map may contain very few samples, and so
their variances can be widely inaccurate. So exclude pixels with low hits
from the estimate of the normalised map change (which is the basis of the
convergence test). "low hits" is anything under 10% of the mean hits per
pixel across the map.