smurf: Correct mapbounds fast-mode so that it can be used for JSA maps
The previous version assumed that north was upwards, which it isn't for
JSA. For instance, try

% makemap in=/star/share/smurf/s4a\* crota=45 \
          config=/star/share/smurf/dimmconfig.lis

The issue was that the extrema time slices were found by looking at the
bounding box in (ra,dec) offset, but if the pixel axes are not aligned
with (Ra,Dec), then that's not the right box to use. Now, the boresight
positions are transformed into output map pixel coords, and the bounding
box in pixel coords is used to find the extrema time slices.

In addition, there may be some time gain due to the fact that the new
scheme uses astTran2 to transform all time slice positions simultaneously,
rather than making a separate call to astTran2 for each time slice.