smurf: Allow holes in time-slices to be filled before forming common-mode
Simply excluding holes in a time slice from the mean value will tend to
bias the mean value if there is a strong gradient across the time slice.
E.g if there is a hole in the bright area, then the mean value will be
biased downwards.

So instead of simply excluding holes, we now have an option to fill holes
with values that are representative of the bolometer values in the
neighbourhood of the holes. This should help to prevent the above bias.

This option is switched on by setting com.fill=1