kaplibs: Fix bug calculating the collapsed variance for the integ estimator.
It failed to square the width in the variance summation.  In many
cases for heterodyne reductions the bin width is 1 km/s, and the bug
would have no effect.  Also the errors in the integ map might not be
used or studied too closely.

Kudos to Sarah for spotting this howler empirically, confirmed using a
dimensional analysis.  Others may have commented that the variances
looked strange but didn't pinpoint the cause.