smurf: Add optional Spectral Filter Profile calibration file handling
A new 'SFP' optional calibration file parameter has been added for this purpose.
The SFP data is in a 1D array of 32 x 40 x nSfp values,
with an expected resolution of 0.025 (1/cm), normalized to 1.
For the 850 band, this ranges from wave number 11.22 to 12.395 in 1/cm units, and
for the 450 band, it ranges from wave number   21.63 to 23.105 in 1/cm units
This data is interpolated to the resolution of the spectrum before dividing it out.

Add optional RESOLUTION override
If defined as a command line parameter, an arbitrary output resolution is applied.

Add optional FNYQUIST override
Revised STEPTIME value to produce a fNyqust of 50.0 at nominal SCANVEL of 4.2 mm/sec