KAPPA: Appease C-shell if-then-else-if logic testing.
It was failing if a non-existent NDF is supplied on the command line
and then attempted to look for a foreign-format file.  The error was
different depending on whether CONVERT definitions had been assigned.

Well I recall giving these UI modifications a thorough test of the
various options in Commit c42130297076e when I introduced this feature
for Jan.  It seems that

if ( expr1 ) then
   <report error>
else if ( expr2 ) then
   <report error>
else
   <do stuff>
endif

needs expr2 to be valid even when expr1 is true.  I've changed
expr1 and expr2.