KAPPA: Do not set the BAD_PIXEL flag in COPYBAD.
The statistic returned from kps1_cpb.gen is the number of bad values
copied.  It has no knowledge of any existing bad values and these may
be different for the DATA and VARAINCE components.  As previously
written it was possible to have the BAD_PIXEL flag set to FALSE while
there were bad values present.

Also change the description of the kps1_cpb.gen arguments, and
COPYBAD's output parameters and messages to match the actual meanings.