smurf: Keep OBSGEO headers from the end of group
This commit adds the OBSGEO-X, OBSGEO-Y and OBSGEO-Z headers to the
list of headers which the smf_fits_outhdr function retains from the
end of the group of input headers if they differ.  This is to prevent
these headers disappearing when merging headers from different
observations between which the position of the telescope changed.