perlmods: Update JSA and UKIRT::Archive modules
Both modules should now write CADC file names in lower case,
as requested by CADC at this week's teleconference.

Provenance processing was refactored a bit in the JSA module to
allow it to be used by the UKIRT::Archive module.  There are also
updates related to the co-adding of JSA tiles.

The UKIRT::Archive module gained provenance processing and handling
of the log.group file.  Data processing date and ID are added to
PNGs, using a subroutine now exported by the JSA module.