TSP: Enable creation of the hypertext documentation.
Removed what appears to be superfluous manitemize environment for
parameter descriptions that makes latex2html run forever.

This still leaves the Parameters not appearing correctly in the html
(they are fine in the PDF).  The output is "@ss@hl".  These are in
TeX-style definitions which are known to cause problems, cf. SUN/199.
At least it's the same problem for all the Parameters sections.