smurf: Allow low frequencies to be removed from map changes
The old experimental "ast.gaussbg" parameter has been replaced by
"ast.dilt_diff".

If ast.filt_diff is non-zero, the map created at the end of each
iteration is modified by removing any large scale structures that have
been introduced by the iteration. Such structures are a natural outcome
of the iterative process and often take the form of ripples on the scale
of the FLT filter size that get stronger with each iteration.

If "ast.filt_diff" is non-zero it gives the size of the largest features
to be retained, in arc-sconds. At the end of each iteration the
difference is taken between the new map and the previous map. This
difference is filtered using a sharp-edged high-pass filter, of size
specified by "ast.filt_diff". The remaining high frequencies (i.e. small
structures that have been introduced by the new iteration) are added back
onto the map created by the previous iteration. This modified map is then
used, instead of the original map, to define the AST model.